refactor(daemon): 主程式搬進 sw_core.daemon，root serialwrapd.py 轉 shim
- 新增 sw_core/daemon.py：原 serialwrapd.py 全部實作（含 BLOCKING_RPC_METHODS、
  build_parser、_run_async、main）搬入套件，改用 package-absolute import
- serialwrapd.py 改寫為薄 shim：re-export BLOCKING_RPC_METHODS 與 main
  維持回溯相容（既有 test_issue52_rpc_concurrency 不須修改）
- 新增 tests/test_packaging_entrypoints.py：TDD 紅綠循環驗證
  sw_core.daemon.main 可呼叫且 BLOCKING_RPC_METHODS 包含 file.push/file.pull
